bin/requirements.txt
jsonschema==4.0.1
PyGithub==1.55
